fix to AIR 3.2 potential bug on iPad about stage dimensions
when running on my iPad stage.stageWidth and stage.stageHeight where
reading 500x375 pixels - I tried everything but what seems to fix it is
using stage.fullScreenWidth and height and compiling the application in
fullscreen mode. This might give inaccurate alignment when running the
application in the browser so you guys may want to revert this. Keep in
mind that if you do the 2D and 3D layout in my iPad is completely
messed up.